Refactor analyze.py to analyzer.py.
