+++ mist/main.py
@@ -1,3 +1,13 @@
+# Copyright (c) MIST Imaging LLC.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#     http://www.apache.org/licenses/LICENSE-2.0
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 """Main script for MIST."""
 import os
 import argparse
@@ -5,7 +15,7 @@ import numpy as np
 import torch
 
 # Import MIST modules.
-from mist.analyze_data.analyze import Analyzer
+from mist.analyze_data.analyzer import Analyzer
 from mist.preprocess_data.preprocess import preprocess_dataset
 from mist.runtime import args
 from mist.runtime.run import Trainer
